Implement Drive -> Get Drives
This fixes #14.

First implementation of a real method.
We also decided to add a 'root' to this, because otherwise you wouldn't
be able to figure out what you're MyDrive is. Hopefully this works when
you combine it with the Folder -> Get List method

Test coverage on the FILE service object is down as that code is no
longer executed in the test set.